backend/mlservice/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.2.0
prophet==1.1.5
mysql-connector-python==8.2.0
python-multipart==0.0.6
redis==5.0.1
aioredis==2.0.1
gunicorn==21.2.0
httpx==0.25.2
numpy==1.26.4
redis==5.0.1
fastapi-cache2==0.2.1
aioredis==2.0.1
azure-ai-documentintelligence==1.0.0b4
azure-identity==1.15.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7
python-dotenv==1.0.0